OrderedConcurrentBuffer: change buckets to be ConcurrentLinkedQueue
It was ConcurrentHashMap<BucketNumber, Bucket> before as I assumed it
can be sparse: remaining of a previous version where buckets were
initialized lazily. I've abandoned the laziness (!, my grandma would be
proud :D), but the idea that buckets can be sparse somehow remained in
my head (thanks to Kamil for enlightenment).
This significantly simplifies the code: 1 AtomicInt and 1 volatile
fields could be completely removed and number of races was reduced from
4 to 2.